[FIX] im_livechat: pin "chat" channels on message post
Revision on https://github.com/odoo/odoo/commit/1c85bcb478a0435d009ecf2757b4137c13ccb5bc

Commit above fixed an issue where livechat channels were not pinned
on the UI after page refresh. To do so, it marked channels of type
"chat" and "livechat" as chat channels.

However, this failed marking channels of type "chat" as chat
channels, because it mistakenly unset these channels while
processing livechat channels. As a result, chat channels were not
marked as pinned, resulting in this channel not being visible in
the interface.

Steps to reproduce:

1. Connect as Admin.
2. Have unpinned conversation with Demo.
3. Receive a new private message from Demo.
4. Refresh page
=> the conversation is not visible in Discuss app or in systray
   messaging menu.

This commit fixes the issue by not overriding channels that have not
been marked as `is_chat`. This way, channels of type "chat" and
"livechat" are correctly considered as chat channels.

closes odoo/odoo#46266

X-original-commit: 72a9498ac93d15767f221e3db9b94f38f1c2e8b5
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>