[FIX] web_editor: fix pictograms in rte
The content of the "wysiwyg.fonts" module is duplicated in
"web_editor.base". This is a mistake that was introduced with the
revert of the saas-12.2 editor.

This duplication is the cause of github issue https://github.com/odoo/odoo/issues/44689.

This commit fixes the issue by requiring the correct module in rte.

closes odoo/odoo#46865

X-original-commit: 8bc3fd2c41c1d0fc1fa22e0849a7204b6712d7b2
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>