[FIX] website_blog: fix s_latest_post using undefined theme color
In odoo/odoo#42466 , the s_latest_posts snippet options that were
theme-specific were brought into standard from various themes. One of
those options is hover effects, and one of those hover effects used the
beta theme color, which is not defined in theme_bootswatch, and would
cause assets_frontend to fail to build when that theme was selected.

This commit fixes that by changing it to use the secondary theme-color
instead, which is available in all themes.

closes odoo/odoo#46387

X-original-commit: 0fae8923015f2abc95c97134a6537c0cffa60596
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>