[FIX] sale: Traceback when creating a quote
The field 'type' in the account.move model has been renamed into 'move_type'
but the ir.rules of the sale module have not been updated.

The fix consists in renaming this field in ir.rules.

As another bug was due to this field renaming (see b90dc8ce2b8), I have checked
that the old 'type' field is no more used in Odoo files.

closes odoo/odoo#46398

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>