[FIX] point_of_sale: Connect to the IoT Box in debug
When using the PoS in debug mode in 12.0, the call to the
'/hw_proxy/handshake' route contains the 'X-Debug-Mode' header.

This header was removed from the 'Access-Control-Allow-Headers' in
13.0, which is used to create the image of the IoT Box.

This means that if the code of http.py was not checked out to 12.0
(i.e. when using an IoT Box in Community), the connection would fail
and none of the devices could be used.

closes odoo/odoo#46442

X-original-commit: e6cfbd35e0878c9c07a98f950495a682c6cff535
Signed-off-by: Antoine Prieëls <aprieels@users.noreply.github.com>