[FIX] account: extract code into method to allow changing the date of the tax
Before this commit, when two lines belonging to a period that is already
closed a cash basis entry is created at the present date. There are
localizations that could not allow creating the cash basis entry in a
date other than the maximum between the Journal Items being reconciled.

Now, the localizations are able to override and chose the date without
having to override the whole 'create_tax_cash_basis_entry' method.

opw-2195016

closes odoo/odoo#46661

X-original-commit: 96f3a5459ed281851a3bee6e86e06caa7be2be06
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>