[IMP] website_slides: implement archive cascade behavior in toggle_active
Purpose of this commit is to make use of toggle_active to implement business
behavior linked to active field being changed. In this commit we move the
business specific code from write to toggle_active. When archiving a channel
its slides are archived.

In this commit we also track active field for courses and slides, allowing
to see flag changes in chatter.

Task ID 2170708
Community PR odoo/odoo#46563