[IMP] web_editor, website: improve border and border radius options
Previously, the border radius option was only available when there was a
border width, but in some cases you want no border, a background color,
and still have rounded corners. This commit makes the option available
when there is either a border OR a background color.

This commit also adds an option to choose the border style between
solid, dashed, dotted and double. The other styles have been left out as
they look very different browser to browser and are not very useful.

task-2195429

closes odoo/odoo#46625

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>