[FIX] base_iban: fix display of IBAN validation
Before this commit, the IBAN number was showing behind the X or V of the
validator.

This commit adds a class on the input when the validator is added, so
that a padding is set on it. The result is that the IBAN text doesn't
show behind the validator any more, once it has been displayed.

closes odoo/odoo#46859

X-original-commit: 0b2c40dbe0df2802b6a1d2969b4e6651a750f928
Signed-off-by: Toufik Benjaa (tbe) <tbe@odoo.com>